1.0.0-beta-sun-build21 Version:
* Addition of support for database sequences
* Code review
* Fix for data type in sequences

+++ CatFactory/ObjectRelationalMapping/ByteSequence.cs
@@ -43,27 +43,27 @@
         /// <summary>
         /// Gets or sets the start value for current sequence
         /// </summary>
-        public short StartValue { get; set; }
+        public byte StartValue { get; set; }
 
         /// <summary>
         /// Gets or sets the increment for current sequence
         /// </summary>
-        public short Increment { get; set; }
+        public byte Increment { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum value for current sequence
         /// </summary>
-        public short MinimumValue { get; set; }
+        public byte MinimumValue { get; set; }
 
         /// <summary>
         /// Gets or sets the maximum value for current sequence
         /// </summary>
-        public short MaximumValue { get; set; }
+        public byte MaximumValue { get; set; }
 
         /// <summary>
         /// Gets or sets the current value for current sequence
         /// </summary>
-        public short CurrentValue { get; set; }
+        public byte CurrentValue { get; set; }
 
         #endregion
     }
