+++ CatFactory/ObjectRelationalMapping/Int32Sequence.cs
@@ -43,27 +43,27 @@
         /// <summary>
         /// Gets or sets the start value for current sequence
         /// </summary>
-        public short StartValue { get; set; }
+        public int StartValue { get; set; }
 
         /// <summary>
         /// Gets or sets the increment for current sequence
         /// </summary>
-        public short Increment { get; set; }
+        public int Increment { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum value for current sequence
         /// </summary>
-        public short MinimumValue { get; set; }
+        public int MinimumValue { get; set; }
 
         /// <summary>
         /// Gets or sets the maximum value for current sequence
         /// </summary>
-        public short MaximumValue { get; set; }
+        public int MaximumValue { get; set; }
 
         /// <summary>
         /// Gets or sets the current value for current sequence
         /// </summary>
-        public short CurrentValue { get; set; }
+        public int CurrentValue { get; set; }
 
         #endregion
     }
