+++ CatFactory/ObjectRelationalMapping/Int64Sequence.cs
@@ -43,27 +43,27 @@
         /// <summary>
         /// Gets or sets the start value for current sequence
         /// </summary>
-        public short StartValue { get; set; }
+        public long StartValue { get; set; }
 
         /// <summary>
         /// Gets or sets the increment for current sequence
         /// </summary>
-        public short Increment { get; set; }
+        public long Increment { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum value for current sequence
         /// </summary>
-        public short MinimumValue { get; set; }
+        public long MinimumValue { get; set; }
 
         /// <summary>
         /// Gets or sets the maximum value for current sequence
         /// </summary>
-        public short MaximumValue { get; set; }
+        public long MaximumValue { get; set; }
 
         /// <summary>
         /// Gets or sets the current value for current sequence
         /// </summary>
-        public short CurrentValue { get; set; }
+        public long CurrentValue { get; set; }
 
         #endregion
     }
